+++ src/main/java/be/baur/sds/content/IntegerType.java
@@ -7,8 +7,11 @@ import java.util.function.Function;
  */
 public final class IntegerType extends RangedType<Integer> {
 
+	/** The SDS name of this data type. */
 	public static final String NAME = "integer";
 
+	/** A function that constructs an integer value from a string. */
+	public static final Function<String, Integer> VALUE_CONSTRUCTOR = Integer::new;
 
 	/**
 	 * Creates the type with the specified name.
@@ -27,14 +30,26 @@ public final class IntegerType extends RangedType<Integer> {
 	}
 
 
-	@Override
-	public Class<Integer> valueClass() {
-		return Integer.class;
-	}
+//	@Override
+//	public Class<Integer> valueClass() {
+//		return Integer.class;
+//	}
 
 
 	@Override
 	public Function<String, Integer> valueConstructor() {
 		return Integer::new;
+	}
+	
+	
+	/**
+	 * Returns an Integer if the supplied string is within the lexical space of this type.
+	 * 
+	 * @param s the string to be converted
+	 * @return an Integer
+	 * @throws NumberFormatException if conversion is not possible
+	 */
+	public static Integer valueOf(String s) {
+		return VALUE_CONSTRUCTOR.apply(s);
 	}
 }